requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-gemini

PyGithub>=2.3.0
requests>=2.31.0

pydantic>=2.8.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

tree-sitter>=0.22.0
tree-sitter-python>=0.22.0
tree-sitter-java
tree-sitter-javascript>=0.21.0

# Utilidades
rich>=13.7.0
typer>=0.12.0
asyncio-throttle>=1.0.2

pytest>=8.0.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0

fastapi
logging
docx
fitz
pinecone
lanchain